+++ coopercars-app/src/components/AddVehicle.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+
+
+function AddVehicle(){
+
+    return(
+        <div>
+        AddVehicle
+        </div>
+    );
+}
+
+export default AddVehicle;
