+++ coopercars-app/src/components/BrowseVehicle.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+
+
+function BrowseVehicle(){
+
+    return(
+        <div>
+        BrowseVehicle
+        </div>
+    );
+}
+
+export default BrowseVehicle;
