+++ coopercars-app/src/components/HomePage.js
@@ -1,49 +1,152 @@
-import React from 'react'
-import AppBar from '@material-ui/core/AppBar'
-import Toolbar from '@material-ui/core/Toolbar'
-import Typography from '@material-ui/core/Typography'
-import Button from '@material-ui/core/Button';
-import {Container, Tooltip} from "@material-ui/core";
-import logo from "../CooperCars-logos_white.png";
+import React, { useState } from 'react';
+//import useInterval from './useInterval';
+import coopercars1_logo from '../CooperCars-logos.jpeg';
+import coopercars2_logo from '../CooperCars-logos_black.png';
+import '../App.css';
+import NavBar from './NavBar'
+import TextField from "@material-ui/core/TextField";
+import {
+  BrowserRouter as Router,
+  Route,
+  Link,
+  Navigate,
+  Routes,
+} from "react-router-dom";
 
-const pages = ['Browse Vehicles', 'Add Vehicles'];
-const settings = ['Profile', 'Account', 'Dashboard', 'Logout'];
-
-const NavBar = () => {
-    return(
-        <div>
-            <AppBar position="static" >
-                <Container maxWidth="xl">
-                <Toolbar disableGutters sx={{ height: 100 }}>
-                    <Typography
-                        variant="h6"
-                        noWrap
-                        component="div"
-                        sx={{ mr: 2, display: { xs: 'none', md: 'flex' } }}
-                    >
-                        <img src={logo} className="App-logo" alt="logo" sx={{ p: 0 }} width = "100" height = "100"/>
-                    </Typography>
-
-                    {/*<div>*/}
-                    {/*    <Link>*/}
-                    {/*        Home*/}
-                    {/*    </Link>*/}
-                    {/*    <Link>*/}
-                    {/*        Search*/}
-                    {/*    </Link>*/}
-                    {/*</div>*/}
-                    {/*<Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>*/}
-                    {/*    Vehicles*/}
-                    {/*</Typography>*/}
-                    <Button color="inherit">Browse Vehicles</Button>
-                    <Button color="inherit">Add Vehicle</Button>
-                    <Tooltip>
-                        <Button color="inherit" align = "right" sx={{ p: 0 }}>Login</Button>
-                    </Tooltip>
-                </Toolbar>
-                </Container>
-            </AppBar>
-        </div>
-    )
+function getRandomColor() {
+  let colorValues = ["red", "blue", "green"];
+  return colorValues[Math.floor(Math.random() * colorValues.length)];
 }
-export default NavBar;
+
+function HomePage() {
+  const apiUrlPrefix = "http://localhost:8080";
+  //const vins = ['1FMCU9GD1HUA30879', '3FA6P0LU1KR242602', '3HGGK5H88KM742051', '5YJ3E1EA5JF098290', '1FTEX1E51KKC66386', '1C4RJFLG1HC603078', '3FA6P0LU1KR101755', 'JTMRWRFV7LJ048851', '1G1RE6E42EU111830', 'SADCJ2BN5HA086947', 'KNMAT2MV3JP608780', '5NPD84LF6KH490922', '5npe34af3jh646547', '2hgfc2f78jh564740', '5YFEPRAE7LP054292', '55SWF4KB5GU142000'];
+
+  const [currentVIN, setCurrentVIN] = useState('-');
+  const [currentVehicleInfo, setCurrentVehicleInfo] = useState('-');
+  const [currentVehicleFeatures1, setCurrentVehicleFeatures1] = useState('-');
+  const [currentVehicleFeatures2, setCurrentVehicleFeatures2] = useState('-');
+  const [currentVehicleFeatures3, setCurrentVehicleFeatures3] = useState('-');
+  const [currentVehicleFeatures4, setCurrentVehicleFeatures4] = useState('-');
+  //const [data,setData] = useState(null);
+  //const [print,setPrint] = useState(false);
+  //const vinDelay = 30000; // in milliseconds
+  const [vin, setVin] = useState("");
+
+
+  HomePage.refreshVehicleInfo = (vin) => {
+    console.log("Refreshing ... %s vehicle ...", currentVIN);
+    // var vin = vins[Math.floor(Math.random() * vins.length)];
+    setCurrentVIN(vin);
+    setCurrentVehicleInfo(vin);
+    setCurrentVehicleFeatures1(vin);
+    setCurrentVehicleFeatures2(vin);
+    setCurrentVehicleFeatures3(vin);
+    setCurrentVehicleFeatures4(vin);
+    var vehicleApiUrl = apiUrlPrefix.concat("/api/vehicle/",vin);
+      fetch(vehicleApiUrl)
+      .then(response => response.json())
+      .then(data => {
+
+        // the following is an example for presentation purposes... we ultimately want to present this in a better way than just sentences
+        setCurrentVehicleInfo(data.year + " " + data.make + " " + data.model + " " + data.trim + " " + data.series);
+        var hasFCW = "";
+        var hasBSW = "";
+        var hasACC = "";
+        var hasBackup = "";
+
+        if (data.forwColliWarn === "Standard"){
+          hasFCW = "- Is equipped "
+        }
+        else{
+          hasFCW = "- Is not equipped "
+        }
+
+        if (data.blindSpotWarn == "Standard"){
+          hasBSW = "- Is equipped "
+        }
+        else{
+          hasBSW = "- Is not equipped "
+        }
+
+        if (data.adaptiveCruiseControl == "Standard"){
+          hasACC = "- Is equipped "
+        }
+        else{
+          hasACC = "- Is not equipped "
+        }
+
+        if (data.backupCam == "Standard"){
+          hasBackup = "- Is equipped "
+        }
+        else{
+          hasBackup = "- Is not equipped "
+        }
+
+
+        setCurrentVehicleFeatures1(hasFCW + "with forward collision warning. ");
+        setCurrentVehicleFeatures2(hasBSW + "with blind spot warning. ");
+        setCurrentVehicleFeatures3(hasACC + "with adaptive cruise control. ");
+        setCurrentVehicleFeatures4(hasBackup + "with a backup camera. ");
+        console.log(data);
+      })
+      .catch(err => {
+        console.log("Cannot connect to API endpoint: %s", vehicleApiUrl);
+      });
+    console.log("Refreshed %s VIN.", currentVIN);
+  }
+  // useInterval(() => {
+  //   HomePage.refreshVehicleInfo();
+  // }, vinDelay);
+
+
+  HomePage.buttonClicked = () => {
+    console.log('Button was clicked!');
+    HomePage.refreshVehicleInfo(vin);
+  }
+
+  // HomePage.getData = (val) => {
+  //     setData(val.target.value)
+  //     console.warn(val.target.value)
+  // }
+
+  return (
+    <div className="App">
+      <NavBar />
+      <header className="App-header">
+      <img src={coopercars1_logo} className="App-logo" alt="cooper-logo" />
+
+        <p>
+          Enter VIN of vehicle to add to inventory:
+        </p>
+
+        <TextField
+            style={{background: "rgb(232, 241, 250)"}}
+            value={vin}
+            onChange={(e) => {
+              setVin(e.target.value);
+            }}
+
+        />
+
+        <button className="button" onClick={HomePage.buttonClicked}>Submit</button>
+
+
+        {/*<input type="text" onChange={HomePage.getData}/>*/}
+        {/*<button onClick={HomePage.getData}>Submit</button>*/}
+
+        <br></br>
+        {/*<button className="button" onClick={HomePage.buttonClicked}>Click to Refresh</button>*/}
+
+        <p>The vehicle for VIN {vin} is </p>
+        {currentVehicleInfo}<br></br>
+        {currentVehicleFeatures1}<br></br>
+        {currentVehicleFeatures2}<br></br>
+        {currentVehicleFeatures3}<br></br>
+        {currentVehicleFeatures4}<br></br>
+      </header>
+    </div>
+  );
+}
+
+export default HomePage;
