+++ coopercars-app/src/components/Login.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+
+
+function Login(){
+
+    return(
+        <div>
+        Login
+        </div>
+    );
+}
+
+export default Login;
